refactor: move utility to test utils
