+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import logging
 import os
 import uuid
-from typing import List
+from typing import List, Union, cast
 
 import numpy as np
 import pandas as pd
@@ -29,7 +29,7 @@ from great_expectations.data_context.store import (
 )
 from great_expectations.data_context.store.util import (
     build_checkpoint_store_using_store_backend,
-    build_profiler_store_using_store_backend,
+    build_configuration_store,
     delete_checkpoint_config_from_store_backend,
     delete_config_from_store_backend,
     load_checkpoint_config_from_store_backend,
@@ -287,6 +287,23 @@ def build_checkpoint_store_using_filesystem(
         store_name=store_name,
         store_backend=store_backend_obj,
         overwrite_existing=overwrite_existing,
+    )
+
+
+def build_profiler_store_using_store_backend(
+    store_name: str,
+    store_backend: Union[StoreBackend, dict],
+    overwrite_existing: bool = False,
+) -> ProfilerStore:
+    return cast(
+        ProfilerStore,
+        build_configuration_store(
+            class_name="ProfilerStore",
+            module_name="great_expectations.data_context.store",
+            store_name=store_name,
+            store_backend=store_backend,
+            overwrite_existing=overwrite_existing,
+        ),
     )
